Merger a Developer
